introducing AccountRepository class
